+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * New cop `RedundantReturn` tracks redundant `return`s in method bodies
 * New cop `RedundantBegin` tracks redundant `begin` blocks in method definitions.
+* New cop `RedundantSelf` tracks redundant uses of `self`.
 
 ### Changes